+++ userView.py
@@ -11,6 +11,8 @@ class UserView(MethodView):
     def post(self):
         if not request.json:
             abort(400)
+        if not User.objects(user_id=request.json['user_id']).first == None:
+            return jsonify({'Result':'User ID already exists'})
         user = User(
             user_id= request.json['user_id'],
             previous_cliqs= request.json['previous_cliqs'],
@@ -27,14 +29,17 @@ class UserView(MethodView):
     @user_view.route('/users/<id>', methods=['GET'])
     def get(id):
         user = User.objects(user_id=id)
+        if user.first() == None:
+            return jsonify({'Result':'User does not exist'})
         return user.to_json()
 
     @user_view.route('/users/<id>', methods=['PUT'])
     def put(id):
-        user = User.objects(user_id=id)
-        print id
-        if user == None:
+        if not request.json:
             abort(400)
+        user = User.objects(user_id=id)
+        if user.first() == None:
+            return jsonify({'Result':'User does not exist'})
         if 'previous_cliqs' in request.json:
             user.update(previous_cliqs, request.json['previous_cliqs'])
         if 'first_name' in request.json:
@@ -52,6 +57,8 @@ class UserView(MethodView):
     @user_view.route('/users/<id>', methods=['DELETE'])
     def delete(id):
         user = User.objects(user_id=id)
+        if user.first() == None:
+            return jsonify({'Result':'User does not exist'})
         user.delete()
         return jsonify ({'Result': 'Deleted'})
